+++ packages/shared/common/defaultSetting.ts
@@ -35,6 +35,8 @@ const defaultSetting: AnyListen.AppSetting = {
   'player.waitPlayEndStopTime': '',
   'player.autoSkipOnError': true,
   'player.isAutoCleanPlayedList': false,
+  'player.isShowStatusBarLyric': false,
+  'player.isShowTitleLyric': false,
   'player.soundEffect.convolution.fileName': '',
   'player.soundEffect.convolution.mainGain': 10,
   'player.soundEffect.convolution.sendGain': 0,