+++ packages/shared/web/lyric-font-player/utils.js
@@ -1,8 +1,12 @@
-
-export const getNow = typeof performance == 'object' && window.performance.now ? window.performance.now.bind(window.performance) : Date.now.bind(Date)
+export const getNow =
+  typeof performance == 'object' && window.performance.now ? window.performance.now.bind(window.performance) : Date.now.bind(Date)
 
 export class TimeoutTools {
   constructor(thresholdTime = 80) {
+    this.setTimeout = window.setTimeout.bind(window)
+    this.clearTimeout = window.clearTimeout.bind(window)
+    this.nextTick = window.requestAnimationFrame.bind(window)
+    this.cancelNextTick = window.cancelAnimationFrame.bind(window)
     this.invokeTime = 0
     this.animationFrameId = null
     this.timeoutId = null
@@ -11,17 +15,18 @@ export class TimeoutTools {
   }
 
   run() {
-    this.animationFrameId = window.requestAnimationFrame(() => {
+    this.animationFrameId = this.nextTick(() => {
       this.animationFrameId = null
       let diff = this.invokeTime - getNow()
       // console.log('diff', diff)
       if (diff > 0) {
         if (diff < this.thresholdTime) return this.run()
         // console.log('run timeout', diff, diff - this.thresholdTime)
-        return this.timeoutId = window.setTimeout(() => {
+        this.timeoutId = this.setTimeout(() => {
           this.timeoutId = null
           this.run()
         }, diff - this.thresholdTime)
+        return this.timeoutId
       }
 
       // console.log('diff', diff)
@@ -39,13 +44,12 @@ export class TimeoutTools {
 
   clear() {
     if (this.animationFrameId) {
-      window.cancelAnimationFrame(this.animationFrameId)
+      this.cancelNextTick(this.animationFrameId)
       this.animationFrameId = null
     }
     if (this.timeoutId) {
-      window.clearTimeout(this.timeoutId)
+      this.clearTimeout(this.timeoutId)
       this.timeoutId = null
     }
   }
 }
-